llama_index_kickstart/requirements.txt
langchain-community
llama-index
pypdf
llama-index-llms-huggingface-api
llama-index-embeddings-langchain
python-dotenv